+++ src/components/toast/Toast.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { ToastsContext } from '../../context/Toasts';
+import { ToastsContext } from 'context/Toasts';
 import './Toast.scss';
 
 export const Toast = props => {
