+++ src/views/general/home/Home.scss
@@ -1,7 +1,7 @@
 section {
   height: 100vh;
   width: 100%;
-  background: rgba(0, 0, 0, 0.1) url('../../assets/landing.jpg') center no-repeat;
+  background: rgba(0, 0, 0, 0.1) url('../../../assets/landing.jpg') center no-repeat;
   background-size: cover;
   background-blend-mode: overlay;
   position: relative;
