+++ src/views/general/login/Login.js
@@ -1,30 +1,28 @@
 import React, { useEffect, useState, useContext } from 'react';
-import { withRouter, Link, Redirect } from 'react-router-dom';
+import { withRouter, Redirect, Link } from 'react-router-dom';
 
-import { firebaseApp } from '../../firebase/init';
-import { AuthContext } from '../../context/Auth';
-import { ToastsContext } from '../../context/Toasts';
+import { firebaseApp } from 'firebase/init';
+import { AuthContext } from 'context/Auth';
+import { ToastsContext } from 'context/Toasts';
 
-import Image from '../../assets/login-bg.png';
+import Image from 'assets/login-bg.png';
 
-const SignUp = ({ history }) => {
+const Login = ({ history }) => {
 
   useEffect(() => {
-    document.title = 'Sign Up - TaskForce'
+    document.title = 'Login - TaskForce'
   }, []);
 
-  const { currentUser } = useContext(AuthContext);
-
-  const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [formSubmitted, setFormSubmitted] = useState(false);
 
   const [toasts, setToasts] = useContext(ToastsContext);
 
-  const handleSignUp = (e) => {
+  const handleLogin = (e) => {
 
     e.preventDefault();
-    if (!email || !password || !name) {
+    if (!email || !password) {
       setToasts([
         ...toasts,
         {
@@ -37,33 +35,14 @@ const SignUp = ({ history }) => {
       ]);
       return;
     }
+    setFormSubmitted(true);
 
-    firebaseApp
-      .auth()
-      .createUserWithEmailAndPassword(email, password)
+    firebaseApp.auth().signInWithEmailAndPassword(email, password)
       .then(() => {
-        const user = firebaseApp.auth().currentUser;
-        user
-          .updateProfile({ displayName: name })
-          .then(() => {
-            setToasts([
-              ...toasts,
-              {
-                id: toasts.length,
-                title: 'Oh Yes',
-                message: 'Signed up successfully.',
-                backgroundColor: '#5cb85c',
-                icon: 'checkmark-circle'
-              }
-            ]);
-            history.push('/');
-          })
-          .catch((err) => {
-            throw Error(err);
-          });
+        history.push('/s');
       })
-      .catch((err) => {
-        // console.log(err);
+      .catch(err => {
+        setFormSubmitted(false);
         handleError(err);
       });
   }
@@ -82,8 +61,10 @@ const SignUp = ({ history }) => {
     );
   }
 
+  const { currentUser } = useContext(AuthContext);
+
   if (currentUser) {
-    return <Redirect to="/dashboard" />;
+    return <Redirect to="/s" />;
   }
 
   return (
@@ -94,24 +75,8 @@ const SignUp = ({ history }) => {
         </div>
         <div className="form-wrapper">
           <form className="form">
-            <p>Getting started</p>
-            <h2>Create your account</h2>
-
-            <div className="floating">
-              <input type="text"
-                name="name"
-                id="name"
-                value={name}
-                onChange={(e) => setName(e.target.value)}
-                placeholder="Your Name"
-                className="floating__input"
-                autoComplete="off"
-                spellCheck="false"
-              />
-              <label htmlFor="name" className="floating__label" data-content="Name">
-                <span className="hidden--visually">Name</span>
-              </label>
-            </div>
+            <p>Continue working</p>
+            <h2>Login your account</h2>
 
             <div className="floating">
               <input type="email"
@@ -119,7 +84,7 @@ const SignUp = ({ history }) => {
                 id="email"
                 value={email}
                 onChange={e => setEmail(e.target.value)}
-                placeholder="Email address"
+                placeholder='Email address'
                 className="floating__input"
                 autoComplete="off"
                 spellCheck="false"
@@ -129,14 +94,13 @@ const SignUp = ({ history }) => {
               </label>
             </div>
 
-
             <div className="floating">
               <input type="password"
                 name="password"
                 id="password"
                 value={password}
                 onChange={e => setPassword(e.target.value)}
-                placeholder="Password"
+                placeholder='Password'
                 className="floating__input"
                 autoComplete="off"
               />
@@ -145,11 +109,14 @@ const SignUp = ({ history }) => {
               </label>
             </div>
 
+            <div className="help-block">
+              Forgot password ? <Link to="/reset-password">Reset</Link>
+            </div>
             <div className="form-buttons">
-              <button type="submit" className="button" onClick={(e) => handleSignUp(e)}>Sign Up</button>
+              <button type="submit" className="button" disabled={formSubmitted} onClick={(e) => handleLogin(e)}> {formSubmitted ? 'Logging in..' : 'Login'}</button>
             </div>
             <div className="help-block">
-              Already have an account? <Link to="/login">Sign in</Link>
+              Don't have an account? <Link to="/signup">Sign up</Link>
             </div>
           </form>
         </div>
@@ -158,4 +125,4 @@ const SignUp = ({ history }) => {
   );
 };
 
-export default withRouter(SignUp);
+export default withRouter(Login);
