+++ src/views/secure/board/Board.scss
@@ -65,8 +65,8 @@
     width: 100%;
     overflow-x: auto;
     overflow-y: hidden;
-    max-height: calc(100vh - 130px);
-    min-height: calc(100vh - 130px);
+    max-height: calc(100vh - 145px);
+    min-height: calc(100vh - 145px);
     list-style-type: none;
     margin: 0;
     padding: 0;
